requirements.txt
flask
flask-sqlalchemy
flask-security
email-validator
mysqlclient
flask-marshmallow
flask_WTF
bcrypt
flask-blueprint
